+++ CHANGELOG.md
@@ -18,11 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added the ability to downsample to 10K or 2K resolution for freesurfer runs
+- Added the ability to run AFNI 3dDespike on template-space BOLD data.
 - Added the ability to ingress TotalReadoutTime from epi field map meta-data from the JSON sidecars.
 - Added the ability to use TotalReadoutTime of epi field maps in the calculation of FSL topup distortion correction.
 - Difference method (``-``) for ``CPAC.utils.configuration.Configuration`` instances
 - Calculate reho and alff when timeseries in template space
 - Added new default pipeline that uses FSL-BET for brain extraction. Previous default pipleine is now called default-deprecated.
+- Added ``fail_fast`` configuration setting and CLI flag
 
 ### Changed
 - Added a level of depth to `working` directories to match `log` and `output` directory structure